+++ src/main/java/ea/badge/domain/Role.java
@@ -16,7 +16,7 @@ public class Role {
     private Long id;
     private String name;
 
-    public Role(Integer id, String name) {
+    public Role(Long id, String name) {
         this.id = id;
         this.name = name;
     }
